wip: remove old components
