--- ui/src/pages/RecoverLostPassword/RecoverLostPassword.validations.ts
@@ -1,5 +1,0 @@
-import * as Yup from 'yup';
-
-export const validationSchema = Yup.object({
-  loginOrEmail: Yup.string().required('Required'),
-});
