+++ ui/src/pages/SecretMain/SecretMain.test.tsx
@@ -4,5 +4,5 @@ import { SecretMain } from './SecretMain';
 test('renders text content', () => {
   render(<SecretMain />);
   const header = screen.getByText(/Shhhh, this is a secret place./i);
-  expect(header).toBeInTheDocument();
+  expect(header).toBeVisible();
 });
